requirements.txt
disco-py
GitPython
pyyaml
pytest
pytest-cov